Rename package from ap_move_calibration to ap_move_master_to_library
- Rename package directory to match repository name
- Update pyproject.toml package name and entry points
- Update Makefile targets to use new package name
- Update MANIFEST.in for new package directory
- Update README.md with new package name and install URLs
- Update all Python imports and test docstrings
