+++ MANIFEST.in
@@ -1,4 +1,4 @@
 include LICENSE
 include README.md
-recursive-include ap_move_calibration *.py
+recursive-include ap_move_master_to_library *.py
 recursive-include tests *.py
