+++ ap_move_master_to_library/__main__.py
@@ -1,5 +1,5 @@
 """
-Entry point for python -m ap_move_calibration.
+Entry point for python -m ap_move_master_to_library.
 """
 
 from .move_calibration import main
